Update cluster-template-aws.yaml
Provider fails to determine the correct base OS for the AMI. It defaults to ubuntu-18.04 as the base OS and there is no matching AMI for ubuntu-18.04 + kubernetes version > 1.28.1. Since we installed the latest version of the track, the provider looks for ubuntu-18.04-v1.28.3 and this image does not exist.